@@ -48,11 +48,15 @@ class ModuleTest extends PHPUnit_Framework_TestCase
     public function testAssertListenerIsCorrectlyRegistered()
     {
         $module         = new Module();
-        $mvcEvent       = $this->getMock('Zend\Mvc\MvcEvent');
-        $application    = $this->getMock('Zend\Mvc\Application', array(), array(), '', false);
-        $eventManager   = $this->getMock('Zend\EventManager\EventManagerInterface');
-        $serviceManager = $this->getMock('Zend\ServiceManager\ServiceManager');
-        $corsListener   = $this->getMock('ZfrCors\Mvc\CorsRequestListener', array(), array(), '', false);
+        $mvcEvent       = $this->getMockBuilder('Zend\Mvc\MvcEvent')->getMock();
+        $application    = $this->getMockBuilder('Zend\Mvc\Application', [], [], '', false)
+            ->disableOriginalConstructor()
+            ->getMock();
+        $eventManager   = $this->getMockBuilder('Zend\EventManager\EventManagerInterface')->getMock();
+        $serviceManager = $this->getMockBuilder('Zend\ServiceManager\ServiceManager')->getMock();
+        $corsListener   = $this->getMockBuilder('ZfrCors\Mvc\CorsRequestListener', [], [], '', false)
+            ->disableOriginalConstructor()
+            ->getMock();
 
         $mvcEvent->expects($this->any())->method('getTarget')->will($this->returnValue($application));
         $application->expects($this->any())->method('getEventManager')->will($this->returnValue($eventManager));
